Added basic gun script and edited camera pitch and yaw in third person shooter script
Created a inheritable basic gun script, including variables:
maxAmmo
damage
fireRate
reloadTime

set the gun to set the current ammo to the maxAmmo when the game starts.
created a bool to check if gun is loading

Edited the cinemachine portion of the ThirdPersonShooter script to increase roation speed of the camera.
Added a public inspector (CameraSpeed) to edit the camera speed in the unity inspector instead of editing it in the script.