Added MainMenu UI + PowerUp pickups + Level changes
Added a Main Menu UI to the game to quit and start game
Added pickups to the game to increase damage and heal health
Made Agents spawn a certain distance away from the player
Removed Light in the scenes and added spotlight to Agents to indicate their view range.
Moved GameOver UI from player prefab so that LevelManager will handle it.